// src/routes/screenName.enum.ts
enum ScreenNameEnum {
  //Auth Screens
  SPLASH_SCREEN = 'SPLASH_SCREEN',
  SPLASH_SCREEN2 = 'SPLASH_SCREEN2',
  LOGIN_SCREEN = 'LOGIN_SCREEN',
  SIGNUP_SCREEN = 'SIGNUP_SCREEN',
  OTP_SCREEN = 'OTP_SCREEN',

//tab 
  BOTTOM_TAB = 'BOTTOM_TAB',
  HOME_SCREEN ="HOME_SCREEN",
WALLET_SCREEN ='WALLET_SCREEN',
MY_BOOKING= 'MY_BOOKING',
ACCOUNT_SCREEN ='ACCOUNT_SCREEN',
SEARCH_SCREEN ='SEARCH_SCREEN',
SPA_DETAILS ='SPA_DETAILS',
SELECTED_STAFF ='SELECTED_STAFF',
SELECTED_SPATIME= 'SELECTED_SPATIME',
CART_SCREEN ='CART_SCREEN',
SALONAT_HOME='SALONAT_HOME',



//Profile Section

  EDIT_PROFILE = 'EDIT_PROFILE',
  CHANGE_PASSWORD = 'CHANGE_PASSWORD',

  NOTIFICATION_SCREEN = 'NOTIFICATION_SCREEN',
  PRIVACY_POLICY = 'PRIVACY_POLICY',
  TERM_CONDITION ='TERM_CONDITION',

 // Featurs
NEARBY_YOU ='NEARBY_YOU'

 
  //Tab screen 

// chat







 


}

export default ScreenNameEnum;